+++ cartridge-cli.lua
@@ -11,6 +11,7 @@ local yaml = require('yaml')
 
 local self_name = fio.basename(arg[0])
 
+local _TARANTOOL = _G._TARANTOOL
 local function VERSION()
     if package.search('cartridge-cli.VERSION') then
         return require('cartridge-cli.VERSION')
@@ -680,26 +681,40 @@ end
 
 -- * ---------------- Project-related functions ----------------
 
+local function format_version(major, minor, patch)
+    major = major or 0
+    minor = minor or 0
+    patch = patch or 0
+    return string.format('%s.%s.%s', major, minor, patch)
+end
+
 local function normalize_version(str)
     local patterns = {
+        "(%d+)%.(%d+)%.(%d+)-(%d+)-(%g+)",
+        "(%d+)%.(%d+)%.(%d+)-(%d+)",
+        "(%d+)%.(%d+)%.(%d+)-(%g+)",
         "(%d+)%.(%d+)%.(%d+)",
         "(%d+)%.(%d+)",
         "(%d+)"
     }
 
     for _, pattern in ipairs(patterns) do
-        local major, minor, patch = string.match(str, pattern)
-
+        local major, minor, patch, count, hash = string.match(str, pattern)
         if major ~= nil then
-            minor = minor or '0'
-            patch = patch or '0'
-
-            return {major, minor, patch}
+            local release = '0'
+            if count ~= nil and hash ~= nil then
+                release = string.format('%s-%s', count, hash)
+            elseif count ~= nil then
+                release = tostring(count)
+            elseif hash ~= nil then
+                release = tostring(hash)
+            end
+            return format_version(major, minor, patch), release
         end
     end
 end
 
-local function detect_version(source_dir)
+local function detect_git_version(source_dir)
     if which('git') == nil then
         return nil
     end
@@ -723,16 +738,15 @@ local function detect_version(source_dir)
         return nil
     end
 
-    if normalize_version(version) == nil then
+    local normalized_version = normalize_version(version)
+    if normalized_version == nil then
 
         warn("Detected version '%s' ignored, " ..
               "because it doesn't look like proper " ..
               "version (major.minor.patch)", version)
     end
 
-    version = normalize_version(version)
-
-    return version, release, commit
+    return normalized_version, release, commit
 end
 
 local function find_rockspec(source_dir)
@@ -767,28 +781,27 @@ local function detect_name_release_version(source_dir, raw_name, raw_version)
     end
 
     if raw_version then
-        if not normalize_version(raw_version) then
-            die("Passed version '%s' should be semantic (major.minor.patch)",
+        version, release = normalize_version(raw_version)
+        if version == nil then
+            die("Passed version '%s' should be semantic (major.minor.patch[-count][-commit])",
                 raw_version)
         end
-        version = normalize_version(raw_version)
-        release = '0'
+        release = release or '0'
     else
-        version, release = detect_version(source_dir)
+        version, release = detect_git_version(source_dir)
         if version == nil then
             die("Failed to detect version from project in directory '%s'." ..
                     "Please pass it explicitly via --version", source_dir)
         end
 
-        info("Detected project version: %s-%s",
-                            table.concat(version, '.'), release)
+        info("Detected project version: %s-%s", version, release)
     end
 
     if not fio.path.exists(fio.pathjoin(source_dir, 'init.lua')) then
         die("Application must have `init.lua` in its root directory")
     end
 
-    return name, release, version
+    return name, version, release
 end
 
 -- * ----------- Special filenames ------------
@@ -1050,7 +1063,7 @@ local function get_rock_versions(project_dir)
     return dependencies
 end
 
-local function generate_version_file(source_dir, dest_dir, app_name, app_version)
+local function generate_version_file(dest_dir, app_name, app_version, app_release)
     local version_file, _ = fio.open(
         fio.pathjoin(dest_dir, 'VERSION'),
         {'O_TRUNC', 'O_WRONLY', 'O_CREAT'},
@@ -1076,8 +1089,7 @@ local function generate_version_file(source_dir, dest_dir, app_name, app_version
         version_file:write(string.format('TARANTOOL=%s\n', _TARANTOOL))
     end
 
-    local _, _, app_commit = detect_version(source_dir)
-    version_file:write(string.format("%s=%s-%s\n", app_name, table.concat(app_version, '.'), app_commit or ''))
+    version_file:write(string.format("%s=%s-%s\n", app_name, app_version, app_release))
 
     local rocks_versions, err = get_rock_versions(dest_dir)
     if rocks_versions == nil then
@@ -1396,10 +1408,8 @@ end
 
 -- * ---------------- TAR.GZ packing ----------------
 
-local function pack_tgz(source_dir, dest_dir, name, release, version)
-    local tgz_file_name = string.format(
-        "%s-%s-%s.tar.gz",
-        name, table.concat(version, '.'), release)
+local function pack_tgz(source_dir, dest_dir, name, version, release)
+    local tgz_file_name = string.format("%s-%s-%s.tar.gz", name, version, release)
     tgz_file_name = fio.pathjoin(dest_dir, tgz_file_name)
 
     info("Packing tar.gz file")
@@ -1418,7 +1428,7 @@ local function pack_tgz(source_dir, dest_dir, name, release, version)
 
     form_distribution_dir(source_dir, distribution_dir)
     build_application(distribution_dir)
-    generate_version_file(source_dir, distribution_dir, name, version)
+    generate_version_file(distribution_dir, name, version, release)
 
     if tarantool_is_enterprise() then
         copy_taranool_binaries(distribution_dir)
@@ -1434,7 +1444,7 @@ end
 
 -- * ---------------- ROCK packing ----------------
 
-local function pack_rock(source_dir, dest_dir, name, release, version)
+local function pack_rock(source_dir, dest_dir, name, version, release)
     local tmpdir = fio.tempdir()
     local destdir = fio.pathjoin(
         tmpdir, name)
@@ -1446,7 +1456,7 @@ local function pack_rock(source_dir, dest_dir, name, release, version)
 
     form_distribution_dir(source_dir, destdir)
     build_application(destdir)
-    generate_version_file(source_dir, destdir, name, version)
+    generate_version_file(destdir, name, version, release)
 
     if tarantool_is_enterprise() then
         copy_taranool_binaries(destdir)
@@ -1459,13 +1469,13 @@ local function pack_rock(source_dir, dest_dir, name, release, version)
     if rockspec then
         content = read_file(fio.pathjoin(destdir, rockspec))
         content = string.gsub(content, "(.-version%s-=%s-['\"])(.-)(['\"].*)",
-                '%1' .. string.format('%s-%s', table.concat(version, '.'), release) .. '%3')
+                '%1' .. string.format('%s-%s', version, release) .. '%3')
         if not content then
             die('Rockspec %s is not valid! Version not found!')
         end
     end
 
-    local name_of_rockspec = string.format('%s-%s-%s.rockspec', name, table.concat(version, '.'),
+    local name_of_rockspec = string.format('%s-%s-%s.rockspec', name, version,
                     release)
 
     local new_rockspec = fio.pathjoin(destdir, name_of_rockspec)
@@ -1474,7 +1484,7 @@ local function pack_rock(source_dir, dest_dir, name, release, version)
 
     fio.chdir(destdir)
 
-    local rock_filename = string.format('%s-%s-%s.*.rock', name, table.concat(version, '.'),
+    local rock_filename = string.format('%s-%s-%s.*.rock', name, version,
                                         release)
 
     print(call('tarantoolctl rocks pack %s ', new_rockspec))
@@ -1878,7 +1888,7 @@ local function generate_fileinfo(source_dir)
     return result
 end
 
-local function pack_cpio(source_dir, name, version, opts)
+local function pack_cpio(source_dir, name, version, release, opts)
     -- The resulting CPIO structure should look like it will be
     -- extracted to /
     -- So it contains /usr/share/tarantool/<app>, systemd unit files and tmpfiles conf
@@ -1904,7 +1914,7 @@ local function pack_cpio(source_dir, name, version, opts)
     form_distribution_dir(source_dir, distribution_dir)
 
     build_application(distribution_dir)
-    generate_version_file(source_dir, distribution_dir, name, version)
+    generate_version_file(distribution_dir, name, version, release)
 
     form_systemd_dir(tmpdir, name, opts)
     write_tmpfiles_conf(tmpdir, name)
@@ -1934,18 +1944,26 @@ local function pack_cpio(source_dir, name, version, opts)
     return archive, fileinfo, payloadsize
 end
 
-local function pack_rpm(source_dir, dest_dir, name, release, version, opts)
+local function tarantool_version()
+    return _TARANTOOL:split('-', 1)[1]
+end
+
+local function tarantool_next_major_version()
+    return tostring(_TARANTOOL:split('.', 1)[1] + 1)
+end
+
+local function pack_rpm(source_dir, dest_dir, name, version, release, opts)
     opts = opts or {}
     local rpm_file_name = fio.pathjoin(
         dest_dir,
         string.format(
             "%s-%s-%s.rpm",
-            name, table.concat(version, '.'), release))
+            name, version, release))
 
     info("Packing rpm file")
     local lead = gen_lead(name)
 
-    local cpio, fileinfo, payloadsize = pack_cpio(source_dir, name, version, opts)
+    local cpio, fileinfo, payloadsize = pack_cpio(source_dir, name, version, release, opts)
 
     local create_user_script_rpm = expand(CREATE_USER_SCRIPT, {
         groupadd = '/usr/sbin/groupadd',
@@ -1956,7 +1974,7 @@ local function pack_rpm(source_dir, dest_dir, name, release, version, opts)
 
     local header_tags = {
         {'NAME', 'STRING', name},
-        {'VERSION', 'STRING', table.concat(version, '.')},
+        {'VERSION', 'STRING', version},
         {'RELEASE', 'STRING', release},
         {'SUMMARY', 'STRING', ''},
         {'DESCRIPTION', 'STRING', ''},
@@ -1993,9 +2011,8 @@ local function pack_rpm(source_dir, dest_dir, name, release, version, opts)
         --- See Dependency Tags section of
         --- - https://docs.fedoraproject.org/ro/Fedora_Draft_Documentation/0.1/html/RPM_Guide/ch-package-structure.html
 
-        local major, minor, patch = unpack(normalize_version(_TARANTOOL))
-        local min_version = ('%s.%s.%s'):format(major, minor, patch)
-        local max_version = ('%s'):format(tonumber(major) + 1)
+        local min_version = tarantool_version()
+        local max_version = tarantool_next_major_version()
 
         -- tarantool >= min_version < max_version
         local tarantool_dependency_tags = {
@@ -2047,14 +2064,14 @@ end
 -- data.tar.xz    : package files
 -- control.tar.xz : control files (control, preinst etc.)
 --
-local function form_deb_control_dir(dest_dir, name, release, version)
+local function form_deb_control_dir(dest_dir, name, version, release)
     fio.mktree(dest_dir)
 
     -- control
     local control_filepath = fio.pathjoin(dest_dir, 'control')
     local control_params = {
         name = name,
-        version = ('%s-%s'):format(table.concat(version, '.'), release),
+        version = ('%s-%s'):format(version, release),
         maintainer = 'Tarantool Cartridge Developer',
         arch = 'all',
         desc = ('Tarantool Cartridge app %s'):format(name),
@@ -2063,9 +2080,8 @@ local function form_deb_control_dir(dest_dir, name, release, version)
 
     if not tarantool_is_enterprise() then
         -- Add tarantool dependency
-        local major, minor, patch = unpack(normalize_version(_TARANTOOL))
-        local min_version = ('%s.%s.%s'):format(major, minor, patch)
-        local max_version = ('%s'):format(tonumber(major) + 1)
+        local min_version = tarantool_version()
+        local max_version = tarantool_next_major_version()
 
         control_params.deps = string.format('tarantool (>= %s), tarantool (<< %s)',
                                             min_version, max_version)
@@ -2101,10 +2117,8 @@ local function form_deb_control_dir(dest_dir, name, release, version)
     fio.chmod(postinst_filepath, tonumber('0755', 8))
 end
 
-local function pack_deb(source_dir, dest_dir, name, release, version, opts)
-    local deb_file_name = string.format(
-            "%s-%s-%s.deb",
-            name, table.concat(version, '.'), release)
+local function pack_deb(source_dir, dest_dir, name, version, release, opts)
+    local deb_file_name = string.format("%s-%s-%s.deb", name, version, release)
 
     local tar = which('tar')
 
@@ -2131,7 +2145,7 @@ local function pack_deb(source_dir, dest_dir, name, release, version, opts)
     -- control.tar.xz
     local control_dir = fio.pathjoin(tmpdir, 'control')
     local control_tgz_path = fio.pathjoin(tmpdir, 'control.tar.xz')
-    form_deb_control_dir(control_dir, name, release, version)
+    form_deb_control_dir(control_dir, name, version, release)
 
     local control_data = call(string.format("cd %s && %s -cvJf - .", control_dir, tar))
     write_file(control_tgz_path, control_data)
@@ -2145,7 +2159,7 @@ local function pack_deb(source_dir, dest_dir, name, release, version, opts)
     form_distribution_dir(source_dir, distribution_dir)
 
     build_application(distribution_dir)
-    generate_version_file(source_dir, distribution_dir, name, version)
+    generate_version_file(distribution_dir, name, version, release)
 
     form_systemd_dir(data_dir, name, opts)
     write_tmpfiles_conf(data_dir, name)
@@ -2188,6 +2202,11 @@ local function validate_from_dockerfile(dockerfile_content)
     end
 end
 
+local function tarantool_repo_version()
+    local parts = _TARANTOOL:split('.')
+    return string.format('%s_%s', parts[1], parts[2])
+end
+
 local function construct_dockerfile(filepath, appname, from)
     local expand_params = {
         name = appname,
@@ -2217,8 +2236,7 @@ local function construct_dockerfile(filepath, appname, from)
     else
         expand_params.install_tarantool = DOCKER_INSTALL_OPENSOURCE_TARANTOOL_TEMPLATE
 
-        local major, minor, _ = unpack(normalize_version(_TARANTOOL))
-        expand_params.tarantool_repo_version = string.format('%s_%s', major, minor)
+        expand_params.tarantool_repo_version = tarantool_repo_version()
     end
 
     -- dockerfile tail is expanded separately to prevent errors
@@ -2229,7 +2247,7 @@ local function construct_dockerfile(filepath, appname, from)
     write_file(filepath, dockerfile_content)
 end
 
-local function pack_docker(source_dir, _, name, release, version, opts)
+local function pack_docker(source_dir, _, name, version, release, opts)
     opts = opts or {}
 
     local docker = which('docker')
@@ -2258,7 +2276,7 @@ local function pack_docker(source_dir, _, name, release, version, opts)
     local distribution_dir = fio.pathjoin(tmpdir, name)
 
     form_distribution_dir(source_dir, distribution_dir)
-    generate_version_file(source_dir, distribution_dir, name, version)
+    generate_version_file(distribution_dir, name, version, release)
 
     local dockerfile_path = fio.pathjoin(tmpdir, 'Dockerfile')
     construct_dockerfile(dockerfile_path, name, from)
@@ -2267,7 +2285,7 @@ local function pack_docker(source_dir, _, name, release, version, opts)
     if opts.tag ~= nil then
         image_fullname = opts.tag
     else
-        image_fullname = string.format('%s:%s-%s', name, table.concat(version, '.'), release)
+        image_fullname = string.format('%s:%s-%s', name, version, release)
     end
     info('Building docker image: %s', image_fullname)
 
@@ -2321,7 +2339,7 @@ local function check_if_deprecated_build_flow_is_ised(source_dir)
 end
 
 local function app_pack(args)
-    local name, release, version = detect_name_release_version(args.path, args.name, args.version)
+    local name, version, release = detect_name_release_version(args.path, args.name, args.version)
     local instantiated_unit_template
     if args.instantiated_unit_template then
         instantiated_unit_template = read_file(args.instantiated_unit_template)
@@ -2350,21 +2368,21 @@ local function app_pack(args)
     end
 
     if args.type == 'rpm' then
-        pack_rpm(args.path, '.', name, release, version, {
+        pack_rpm(args.path, '.', name, version, release, {
             unit_template = unit_template,
             instantiated_unit_template = instantiated_unit_template
         })
     elseif args.type == 'deb' then
-        pack_deb(args.path, '.', name, release, version, {
+        pack_deb(args.path, '.', name, version, release, {
             unit_template = unit_template,
             instantiated_unit_template = instantiated_unit_template
         })
     elseif args.type == 'tgz' then
-        pack_tgz(args.path, '.', name, release, version)
+        pack_tgz(args.path, '.', name, version, release)
     elseif args.type == 'rock' then
-        pack_rock(args.path, '.', name, release, version)
+        pack_rock(args.path, '.', name, version, release)
     elseif args.type == 'docker' then
-        pack_docker(args.path, '.', name, release, version, {
+        pack_docker(args.path, '.', name, version, release, {
             tag = args.tag,
             from = args.from,
             download_token = args.download_token,