+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 <!-- Please update cartridge-cli/VERSION.lua with new release -->
 
+### Added
+
+- Allow to pass `--version` in format `major.minor.patch[-count][-hash]`
+
 ## [1.3.0] - 2020-01-13
 
 ### Added